// src/interfaces/staff.ts
export interface IStaff {
  access: { ['param']: any };
  address: string;
  clinic: string;
  created_at: Date;
  d_o_b: string;
  department: string;
  department_route: string;
  designation: string;
  email: string;
  fname: string;
  gender: string;
  hospital_branch_code: string;
  isSuperAdmin: boolean;
  is_online: boolean;
  lname: string;
  online_status: string;
  password: string;
  phone: string;
  smart_code: string;
  status: string;
  updated_at: Date;
  username: string;
  _id: string;
}